==> 0_get_weather_data.ipynb <==
import urllib.request
import pandas as pd
import json
import glob

dfs=[]

dates = pd.date_range('2018-09-27','2018-12-31')

keys = pd.read_csv('../data/keys_darksky.csv')['key']

cities = pd.read_csv('../data/cities_info.csv')[['site_id','location']]

# lat_long = (29.4241,-98.4936)

# for date in dates:
#     time = str(date)[0:10]
#     url = f'https://api.darksky.net/forecast/{key}/{lat_long[0]},{lat_long[1]},{time}T00:00:00Z'
#     url = url + '?lang=en&exclude=currently,minutely,daily,alerts,flags&units=si'
#     print(url)
#     r = urllib.request.urlopen(url)
#     s = r.read()
#     s = json.loads(s.decode("utf-8"))
#     df = pd.DataFrame(s['hourly']['data'])
#     dfs.append(df)
#     print('done')

# df_all = pd.concat(dfs,axis=0)
# df_all = df_all[~df_all.duplicated()]
# df_all['site_id']=9
# df_all.to_csv('../data/darksky/all_sanantonio.csv',index=False)

# Combine all files together

files_list = glob.glob("../data/darksky/*.csv")
files_list = [f for f in files_list if 'all_data' not in f]

df_list = []
for f in files_list:
    df = pd.read_csv(f)
    df['site_id'] = cities[cities['location']==f[20:-4]].index[0]
    df_list.append(df) 
    
df_list = pd.concat(df_list,axis=0)

df_list.to_csv('../data/darksky/all_data.csv',index=False)
