==> 0_UCL_data_scrapping.ipynb <==
import zipfile
import wget
import pandas as pd
import glob
import numpy as np
import re
from sklearn import linear_model
import matplotlib.pyplot as plt

# Get the building ID

df = pd.read_csv('../data/UCL/id_UCL.csv')

# Download data

# ids=[]
# for i in df['id']:
#     for year in [2017,2018]:
#         try : 
#             wget.download(f'https://platform.carbonculture.net/apps/studydata/data-download?year={year}&id={i}&subject=places')
#             ids.append((id,year))
#             print("ID ", i, " Year ", year, " exists")
#         except:
#             print("ID ", i, " Year ", year, " doesn't exist")
#             break

# file_list = glob.glob("./*.zip")

# for file in file_list:
#     with zipfile.ZipFile(file, 'r') as zip_ref:
#         zip_ref.extractall('./extracted/')
#         print(file, "extracted!")

file_list = glob.glob("../data/UCL/UCL_extracted/*.csv")

dfs = []

for file in file_list:
    df = pd.read_csv(file, skiprows=3, index_col=0)
    df = df.stack().reset_index()
    df.columns = ['date','time','meter_reading']
    df['timestamp'] = pd.to_datetime(df['date'] + ' ' + df['time'])
    df['site_id'] = re.findall('\.[0-9]{4}_',file)[0][1:5]
    df['type'] = re.findall('_[a-z]{4}\.csv',file)[0][1:5]
    
    print("File has been read and added", file)
    
    dfs.append(df)

dfs = pd.concat(dfs,axis=0)
dfs['id'] = dfs['site_id'] + '_' + dfs['type']
ucl_data = pd.pivot_table(dfs,index=["timestamp"],columns=['id'], values=['meter_reading'])
ucl_data.columns = [col[1] for col in ucl_data.columns]
ucl_data = ucl_data.resample('1H').sum()

# Aim is to match UCL building to site 1 buildings

train = pd.read_csv('../data/ashrae/train.csv')

d_types = {'building_id': np.int16,
          'meter': np.int8}

for feature in d_types:
    train[feature] = train[feature].astype(d_types[feature])
    
train["timestamp"] = pd.to_datetime(train["timestamp"])

building = pd.read_csv('../data/ashrae/building_metadata.csv')
d_types = {'building_id': np.int16,
          'site_id': np.int8,
          'primary_use': 'category',
          'square_feet': np.int32,
          'year_built': np.float16,
          'floor_count': np.float16}

for feature in d_types:
    building[feature] = building[feature].astype(d_types[feature])
    
building['site_id'] = building['site_id'].replace({8:0,11:7})

train = pd.merge(train,building, how='left')
train = train[train['site_id']==1]
train['id'] = train['building_id'].astype(str) + '_' + train['meter'].astype(str)

train_pt = pd.pivot_table(train,index=["timestamp"],columns=['id'], values=['meter_reading'])
train_pt.columns = [col[1] for col in train_pt.columns]

corr_mat = {}
for site_id in train_pt.columns:
    subset = train_pt[[site_id]].join(ucl_data)
    corr_mat[site_id] = subset.corr()[site_id].iloc[1:]
    
corr_mat = pd.DataFrame(corr_mat)

mapping = pd.DataFrame([])
mapping['UCL_id'] = corr_mat.apply(lambda x: np.argmax(x), axis=0)
mapping['score'] = corr_mat.apply(lambda x: max(x), axis=0)

ucl_data = ucl_data[mapping['UCL_id']]
ucl_data.columns = mapping.index
ucl_data = ucl_data.stack().reset_index()
ucl_data.columns = ['timestamp','level_1','meter_reading']
ucl_data['building_id']=ucl_data['level_1'].str.split('_').apply(lambda x: x[0])
ucl_data['meter']=ucl_data['level_1'].str.split('_').apply(lambda x: x[1])
ucl_data = ucl_data[['timestamp','building_id','meter','meter_reading']]

ucl_data.to_csv('../data/UCL/ucl_data_prepared.csv',index=False)
