==> 1_solar_radiance.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import os, gc, sys, warnings, random, math, psutil, pickle

# Input data files are available in the "../input/" directory.
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

from sklearn.preprocessing import LabelEncoder

warnings.filterwarnings('ignore')

weather_train = pd.read_csv(
    '../../data-vol-1/weather_train.csv'
)

weather_train.timestamp = pd.to_datetime(weather_train.timestamp, format='%Y-%m-%d %H:%M:%S')

weather_test = pd.read_csv(
    "../../data-vol-1/weather_test.csv"
)

weather_test.timestamp = pd.to_datetime(weather_test.timestamp, format='%Y-%m-%d %H:%M:%S')

external_weather = pd.read_csv(
    '../../data-vol-1/more_weather_locations.csv',
    index_col=0,
    parse_dates=True,
    infer_datetime_format=True,
).join(
    pd.read_csv(
        '../../data-vol-1/temperature.csv',
        index_col=0,
        parse_dates=True,
        infer_datetime_format=True,
    ).sub(273.15),
    how='left',
)

external_weather.head()

## [Time Zone](https://www.kaggle.com/datadugong/locate-better-cities-by-weather-temp-fill-nans)

if there is time zone details, then no need to align the timestamp of each site

### Holidays in the city/country


### City

weather = pd.concat([weather_train, weather_test]).set_index(['site_id', 'timestamp'])

weather = weather.reindex(
    pd.MultiIndex.from_product(
        [range(16), pd.date_range('2016-01-01', '2018-12-31 23:00', freq='H')])
)

weather = weather.unstack(level=0).interpolate(limit=2).ffill(limit=1).bfill(limit=1)

w = weather_test.set_index(['site_id', 'timestamp'])
w = w.unstack('site_id')

to_lieaner_interp = []
to_lieaner_interp_idx = []

for col, S in w.iteritems():
    to_lieaner_interp.append(
        S.isnull().astype(int).groupby(S.notnull().astype(int).cumsum()).sum().max()
    )

    to_lieaner_interp_idx.append(col)

pd.DataFrame(to_lieaner_interp, index=pd.MultiIndex.from_tuples(to_lieaner_interp_idx)).unstack(0).rename(columns={0: 'max_consecutive_nans'}, level=0)

### Timezone, Latitude, Longtitude

# in site_id.csv

site_info = weather_train = pd.read_csv('../../data-vol-1/cities_info.csv')
site_info

### sun info

!pip install pysolar

import math
from datetime import timedelta
from pyephem_sunpath.sunpath import sunpos
from datetime import datetime, timedelta
from pysolar.solar import *

df = pd.DataFrame(columns=['site_id','local_time','utc_time','altitude','azimuth','radiation'])

for index, row in site_info.iterrows():
    tz = int(str(row['utc'][:3]))
    site_id = row['site_id']
    lat_long = row['lat_long'].replace('(','').replace(')','').split(',')
    lat = float(lat_long[0])
    long = float(lat_long[1])
    for i in pd.date_range(start='2015-12-31 16:00:00', end='2019-01-01 09:00:00', freq='H'):
        # all output fits in local time      
        u = i - timedelta(hours=tz)
        d = datetime.datetime(i.year, i.month, i.day, i.hour, tzinfo=datetime.timezone.utc)
        alt, azm = sunpos(i, lat, long, tz, dst=False)
        r = radiation.get_radiation_direct(d, alt)
        df = df.append({'site_id':site_id,'local_time':i,'utc_time':u,'altitude':alt,'azimuth':azm,'radiation':r}, ignore_index=True)
    
#to_csv(site_id + '_' + site +'_sun.csv')

df.head()

df = pd.read_csv(r'../../data-vol-1/dates_related/radiance.csv')

df.head()

df.to_csv(r'../../data-vol-1/dates_related/radiance.csv')

Below are backup code

import missingno as msno

msno.matrix(weather_train)

msno.matrix(weather_test)

The missingno correlation heatmap measures nullity correlation: how strongly the presence or absence of one variable affects the presence of another:

a = msno.heatmap(weather_train, sort='ascending')

## Fill NA in weather and missing timestamp

### Wind_Speed
### Wind_direction
### Sea_level_pressure
- no values for site 5
### Precip_depth_1_hr
- no values for site 1, 5, 12
### Dew_temperature
### Cloud_coverage:
- much higher on site 12 than on all other sites
- no values for site 7, 11
### Air_temperature
### Dew temperture

# NOAA Calculation, results is wrong

# fractional year: r
    r = (2*math.pi/365) * (i.timetuple().tm_yday - 1 + (i.hour-12)/24)
    eqtime = 229.18*(0.000075 + 0.001868*math.cos(r) - 0.032077*math.sin(r) - 0.014615*math.cos(2*r) 
                     - 0.040849*math.sin(2*r))
    print('eqtime:' + str(eqtime))
    decl = 0.006918 - 0.399912*math.cos(r) + 0.070257*math.sin(r) - 0.006758*math.cos(2*r) + 0.000907*math.sin(2*r) - 0.002697*math.cos(3*r) + 0.00148*math.sin(3*r)
        
    if ((i >= dateutil.parser.parse(DST_Start)) or i < dateutil.parser.parse(DST_End)):
        utc = summer_UTC
    
    if ((i < dateutil.parser.parse(DST_Start)) or i >= dateutil.parser.parse(DST_End)):
        utc = winter_UTC
        
    time_offset = eqtime + 4*long - 60*utc
    print('time_offset: ' + str(time_offset))
    
    # true solar time
    tst = i.hour*60 + i.minute + i.second/60 + time_offset
    
    # solar hour angle
    ha = (tst/4) - 180
    
    # solar zenith angle
    zenith = math.acos(math.sin(lat*math.pi/180)*math.sin(decl) + math.cos(lat*math.pi/180)*math.cos(decl)*math.cos(ha))
    
    # solar azimuth
    azimuth = 180- math.acos((math.sin(decl)-math.sin(lat*math.pi/180)*math.cos(zenith))/(math.cos(lat*math.pi/180)*math.sin(zenith)))
    
    # sunrise
    sunrise_angle = np.arccos(math.cos(90.833*math.pi/180)/(math.cos(lat*math.pi/180)*math.cos(decl))-math.tan(lat*math.pi/180)*math.tan(decl))
    print('sunrise_angle:' + str(sunrise_angle))
    # sunset
    sunset_angle = -np.arccos(math.cos(90.833*math.pi/180)/(math.cos(lat*math.pi/180)*math.cos(decl))-math.tan(lat*math.pi/180)*math.tan(decl))
    print('sunset_angle:' + str(sunset_angle))
    
    sunrise_utc = 720 - 4*(long + sunrise_angle) - eqtime
    sunset_utc = 720 - 4*(long + sunset_angle) - eqtime
    
    print('sunrise_utc: ' + str(sunrise_utc/60))
    print('sunset_utc: ' + str(sunset_utc/60))
    
    snoon = 720 - 4*long - eqtime
    df = df.append({'datetime':i,'utc':utc,'utc_time':i + timedelta(hours=utc),'r':r,'eqtime':eqtime,'decl':decl,'tst':tst,
                    'ha':ha,'zenith':zenith,'azimuth':azimuth,'sunset_angle':sunset_angle, 
                    'sunrise_angle':sunrise_angle,'sunrise_utc':sunrise_utc/60,
                    'sunset_utc':sunset_utc/60,'snoon':snoon/60},ignore_index = True)
