==> 2_generate_test_leaked.ipynb <==
This is include site0 & site1 & asu & site 4

import warnings
warnings.filterwarnings('ignore')
import os
import gc
import pickle
import numpy as np
import pandas as pd

import feather

# Generate leaked file

Load data

path = '../data'
leaked = feather.read_dataframe(f'{path}/Leaked/leak_isamu.feather')

test = feather.read_dataframe(f'{path}/Prepared_data/test.feather')
test['timestamp'] = pd.to_datetime(test['timestamp'])
test.drop(columns='local_time',inplace=True)

Leaked data 

leaked_all = pd.merge(test, leaked, how='left', on=['building_id','meter','timestamp'])
leaked_all = leaked_all.dropna(subset=['meter_reading'])

leaked_all['row_id'] = leaked_all['row_id'].astype(int)
leaked_all.sort_values('row_id',inplace=True)
leaked_all.reset_index(drop=True,inplace=True)
leaked_all[['row_id','meter_reading']].to_feather('../Leaked/leaked_submission.feather')

# leaked_path = '../data/Leaked'
# leaked_path = '../../data-vol-3/Leaked_data'
# leaked = feather.read_dataframe(f'{leaked_path}/leaked_submission.feather')

# prepared_path = '../../data-vol-3/Prepared_data'
# test = feather.read_dataframe(f'{prepared_path}/test_filled_na.feather')
# test.set_index('row_id',inplace=True)
# test = test.loc[leaked['row_id'].values]
# test['meter_reading'] = leaked['meter_reading']

# test.reset_index().to_feather(f'{prepared_path}/test_leaked_filled_na.feather')

# %%time
# test = feather.read_dataframe(f'../../data-vol-3/Prepared_data/test_leaked_prepared.feather')
# test.drop(columns=['utc','utc_time','altitude','azimuth'], inplace=True)
# test.dropna(subset=['meter_reading'],inplace=True)
# test = test[test['meter_reading']!=0].reset_index(drop=True)
# gc.collect()

# %%time
# test = feather.read_dataframe(f"{path}/Prepared_data/test_prepared.feather")
# test = test.set_index('row_id')
# test = test.sort_index()
# gc.collect()

# leaked = feather.read_dataframe(f"../../data-vol-3/Leaked_data/leaked_submission.feather")
# test = test.iloc[leaked['row_id'].values]
# test['meter_reading'] = leaked['meter_reading']
# test['meter_reading_log'] = np.log1p(test['meter_reading'])
# test.reset_index(inplace=True)

# del leaked
# gc.collect()
# test.to_feather(f'../../data-vol-3/Prepared_data/test_leaked_prepared.feather')
