==> 5. _copy_submission.ipynb <==
import pandas as pd
import zipfile
import numpy as np
import random
import gc
import matplotlib.pyplot as plt
import glob
import csv   



n = '12'
zf = zipfile.ZipFile('../../data-vol-3/Copy_submission/submission_' + n + '.csv.zip') 
df = pd.read_csv(zf.open('submission_' + n + '.csv'))
df.to_feather('../../data-vol-3/Copy_submission/submission_' + n + '.feather')

gc.collect()

df_12 = pd.read_feather('../../data-vol-3/Copy_submission/submission_12.feather')

df_6 = pd.read_feather('../../data-vol-3/Copy_submission/submission_6.feather')

df_2 = pd.read_feather('../../data-vol-3/Copy_submission/submission_2.feather')

df_4 = pd.read_feather('../../data-vol-3/Copy_submission/submission_4.feather')

df_5 = pd.read_feather('../../data-vol-3/Copy_submission/submission_5.feather')

df_7 = pd.read_feather('../../data-vol-3/Copy_submission/submission_7.feather')

df_8 = pd.read_feather('../../data-vol-3/Copy_submission/submission_8.feather')

df_6.head()

df_0['meter_reading'].corr(df_3['meter_reading'])

df_0['meter_reading'].corr(df_8['meter_reading'])

df_0['meter_reading'].corr(df_9['meter_reading'])

df_0['meter_reading'].corr(df_10['meter_reading'])

df_0['meter_reading'].corr(df_11['meter_reading'])

df_0['meter_reading'].corr(df_12['meter_reading'])

(df_0['meter_reading'] + df_3['meter_reading'])/2

df['meter_reading'] = (df_0['meter_reading'] + df_3['meter_reading'] + df_8['meter_reading'] +
                        df_9['meter_reading'] + df_10['meter_reading'] + df_11['meter_reading'] + 
                        df_12['meter_reading'])/7

df.head()

df.to_csv('../../data-vol-3/Copy_submission/submission.csv')
