==> 5_get_submission_all.ipynb <==
# !pip install catboost
# !pip install lightgbm
# !pip install feather-format

import numpy as np
import pandas as pd
from sklearn.model_selection import GroupShuffleSplit
import random
import gc
import lightgbm as lgb
import pickle
import seaborn as sns
import matplotlib.pyplot as plt
import feather
import os
import glob
import csv   
from sklearn.metrics import mean_squared_error

Set seed

seed = 7
random.seed(seed)
np.random.seed(seed)

# Load model

name = '40'

m = f'models/lgbm_v_{name}_fold_all.bin'
estimator = pickle.load(open(m,'rb'))

estimator.params

# Load test data

%%time
path = '../data'
test = feather.read_dataframe(f"{path}/Prepared_data/test_filled_na_light.feather")
gc.collect()

for f in ['primary_use','is_day_saving','is_holiday','building_id','meter','site_id','weekday']:
    test[f] = test[f].astype('category')

#### Reduce memory

test = test[[c for c in test.columns if c not in ['year_built','floor_count','utc','new_year_built',
          'square_feet_log','night_time','morning','afternoon','evening',
          'timestamp','precip_depth_1_hr','wind_speed',
          'new_sea_level_pressure','new_cloud_coverage',
          'new_precip_depth_1_hr','uvIndex','cloudCover','longitude',
          'altitude','azimuth','cloudCover_max_lag72',
          'cloudCover_min_lag72','beaufort_scale','wind_direction_cos','wind_direction_sin',
          'new_wind_direction_sin']]]
gc.collect()

#### Get train data to merge

%%time
print(test.shape)
min_max_hour = feather.read_dataframe(f"{path}/Prepared_data/min_max_by_hour.feather")
test['hour'] = test['utc_time'].dt.hour
test = pd.merge(test, min_max_hour, on=['building_id', 'meter', 'hour'],how='left')
test.drop(columns=['hour','utc_time'],inplace=True)
del min_max_hour
print(test.shape)

%%time
print(test.shape)
min_max = feather.read_dataframe(f"{path}/Prepared_data/min_max.feather")
test = pd.merge(test, min_max, on=['building_id', 'meter'],how='left')
print(test.shape)

# # Save 
# test = test.set_index('row_id')
# test = test.sort_index()
# test.reset_index().to_feather(f'{path}/Prepared_data/test_filled_na_ready.feather')

# test[['building_id','site_id']].reset_index().to_feather(f'{path}/Prepared_data/building_site_mapping.feather')

### Keep only necessary columns

test = test[estimator.feature_name() + ['row_id']]
gc.collect()

# Get results

step = 100000
i = 0
test['pred'] = np.nan

print('here')
while i < test.shape[0]:
    print(i)
    subset = test.loc[i:(i+step)].copy()
    test.loc[i:(i+step),'pred'] = estimator.predict(subset[estimator.feature_name()])
    i = i + step
    gc.collect()

results = test[['pred','row_id']]
results['meter_reading'] = np.exp(results['pred'])-1
results['meter_reading'] = np.where(results['meter_reading'].lt(0),0,results['meter_reading'])
results.sort_values('row_id',inplace=True)

results[['row_id','meter_reading']].reset_index(drop=True).to_feather(f'{path}/submission/submission_v{name}.feather')
results[['row_id','meter_reading']].reset_index(drop=True).to_csv(f'{path}/submission/submission_v{name}.csv', index=False)
